+++ opengb/server.py
@@ -99,7 +99,7 @@ class MessageHandler(object):
         })
         return True
 
-    def move_head_relative(self, x=0, y=0, z=0):
+    def move_head_relative(self, x=0, y=0, z=0, rate=300):
         """
         Move the print head along one or more axes relative to the current
         position.
@@ -110,6 +110,8 @@ class MessageHandler(object):
         :type y: :class:`float`
         :param z: Millimeters to move along the Z axis.
         :type z: :class:`float`
+        :param rate: Rate at which to move in mm/s.
+        :type rate: :class:`float`
         """
         self._to_printer.put({
             'method':   'move_head_relative',
@@ -121,7 +123,7 @@ class MessageHandler(object):
         })
         return True
 
-    def move_head_absolute(self, x=0, y=0, z=0):
+    def move_head_absolute(self, x=0, y=0, z=0, rate=300):
         """
         Move the print head along one or more axes to an absolute position.
 
@@ -131,13 +133,16 @@ class MessageHandler(object):
         :type y: :class:`float`
         :param z: Position to move to along the Z axis.
         :type z: :class:`float`
+        :param rate: Rate at which to move in mm/s.
+        :type rate: :class:`float`
         """
         self._to_printer.put({
             'method':   'move_head_absolute',
             'params': {
-                'x':    x,
-                'y':    y,
-                'z':    z,
+                'x':        x,
+                'y':        y,
+                'z':        z,
+                'rate':     rate,
             }
         })
         return True
@@ -163,7 +168,7 @@ class MessageHandler(object):
         })
         return True
 
-    def retract_filament(self, head, length, rate):
+    def retract_filament(self, head, length, rate=300):
         """
         Retract filament into the print head.
 
@@ -186,7 +191,7 @@ class MessageHandler(object):
         })
         return True
 
-    def unretract_filament(self, head, length, rate):
+    def unretract_filament(self, head, length, rate=300):
         """
         Unretract filament from the print head.